+++ babel.config.js
@@ -8,9 +8,5 @@ module.exports = {
         },
       },
     ],
-  ],
-  plugins: [
-    ["@babel/plugin-proposal-decorators", { legacy: true }],
-    ["@babel/plugin-proposal-class-properties", { loose: true }],
-  ],
+  ]
 };