+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.1.0](https://github.com/aim-leo/tegund/compare/v1.0.3...v2.1.0) (2021-03-12)
+
+
+### Features
+
+* 🎸 add 2.0 fe ([ce3ebe9](https://github.com/aim-leo/tegund/commit/ce3ebe98ae4d721f6995225e41eb66723551b20b))
+* 🎸 add alias for type ([c0b7800](https://github.com/aim-leo/tegund/commit/c0b780083ca20cceb7f864d9b2a11da4d99e6829))
+* 🎸 add array type ([7b1f60d](https://github.com/aim-leo/tegund/commit/7b1f60d7ee94d6934ef70ed2715a9e9aeef9eb4e))
+* 🎸 add basic check ([59a6d19](https://github.com/aim-leo/tegund/commit/59a6d191d684f1a5ce4444a0edde4cda7fd3f760))
+* 🎸 add clone method ([ee0ea4a](https://github.com/aim-leo/tegund/commit/ee0ea4a709bc4a10859a7f849ad5d13b0656eb1e))
+* 🎸 add enum ([d9e7c2a](https://github.com/aim-leo/tegund/commit/d9e7c2a0978064af9862a25d35b626e1d7bfae68))
+* 🎸 add length max min ([2cf3797](https://github.com/aim-leo/tegund/commit/2cf3797fc03d9857667894eef829c2403509cb84))
+* 🎸 add never type ([51fe125](https://github.com/aim-leo/tegund/commit/51fe1255d5979ee0c1a114cd1e6ff06628b1557b))
+* 🎸 add object strict mode && test ([ad0d392](https://github.com/aim-leo/tegund/commit/ad0d392f7b70a4d42e9a45d4ed9c68141314112a))
+* 🎸 add optional ([29e3ca6](https://github.com/aim-leo/tegund/commit/29e3ca63a632a2388debb04f3bb232f6cbfe3a36))
+* 🎸 add partten validator ([f812319](https://github.com/aim-leo/tegund/commit/f8123199c673672889f9863e8311c303cab200e1))
+* 🎸 add range validate mixin ([0d6cc18](https://github.com/aim-leo/tegund/commit/0d6cc184216fb90ce090d2d39a349f9a9b3c6500))
+* 🎸 add ValidateError && inspect ([cc38170](https://github.com/aim-leo/tegund/commit/cc381700fd5cf12e1172fb6fa83d77f12af3523a))
+* 🎸 enable extend multi object ([478ab89](https://github.com/aim-leo/tegund/commit/478ab89e9a64fbf842444ed49a9a5a09fdb88ba3))
+* 🎸 enable set min max of date && add test ([fb3c35e](https://github.com/aim-leo/tegund/commit/fb3c35edc01a14815f5ac35e0aae231ef45aa3dd))
+* add array struct validate ([8fe83f0](https://github.com/aim-leo/tegund/commit/8fe83f0136184b91b936a997281ae877ca85ffb2))
+* add defineProp for MongoT ([f1c8de8](https://github.com/aim-leo/tegund/commit/f1c8de84f76c85108c86cbabecfd41f3883d1e87))
+* add mongund ([9a293e4](https://github.com/aim-leo/tegund/commit/9a293e4d2dfcc7180c8261cc350539596ab63939))
+* add number range mixin ([adaed17](https://github.com/aim-leo/tegund/commit/adaed17e82c6ab2189a0f2c02476cc76599d9fe9))
+* add toModel toSchema function ([cd49e2f](https://github.com/aim-leo/tegund/commit/cd49e2fda61c51a4873e5ea7f91a96e38bdb67c8))
+* pre to write populate ([f6157f0](https://github.com/aim-leo/tegund/commit/f6157f08fa0f5523de37dffc9a3e9cf3f6764849))
+* publish version 2.0.0 ([b24b458](https://github.com/aim-leo/tegund/commit/b24b458753db05572523efd0ea0d687454342331))
+* tmp save ([675c39a](https://github.com/aim-leo/tegund/commit/675c39a3ab31d408401df616f24f96599e1b34bd))
+* tmp save ([79860cc](https://github.com/aim-leo/tegund/commit/79860cc432b4c8c63d8253519efc1f67c19143ec))
+
+
+### Bug Fixes
+
+* 🐛 at condition error ([a6b9bcc](https://github.com/aim-leo/tegund/commit/a6b9bcc7e8d5abab5acbdade2879383dbe1f937c))
+* 🐛 partten -> pattern ([e4d611f](https://github.com/aim-leo/tegund/commit/e4d611fa53b572be505057bd91cef71a035fbc40))
+* 🐛 when check none, should return false ([b860292](https://github.com/aim-leo/tegund/commit/b86029208bd8e4f6c20f1783f2d90bc72b5d3b4f))
+
 ### [1.0.3](https://github.com/aim-leo/tegund/compare/v1.0.2...v1.0.3) (2020-10-11)